Frontend Coupon Feature - Fixing the Live Calculation Issue(Part-3)

+++ resources/views/frontend/layouts/global_scripts.blade.php
@@ -107,7 +107,7 @@
    /** Get current cart total amount **/ 
 
    function getCartTotal(){
-      return parseInt("{{cartTotal()}}");
+      return parseFloat("{{cartTotal()}}");
    }
